requirements.txt
asgiref==3.8.1
Cerberus==1.3.5
Django==5.0.4
djangorestframework==3.15.1
mysqlclient==2.2.4
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
